% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statics.R
\docType{data}
\name{colors_moss}
\alias{colors_moss}
\title{Standardised color set for mossR}
\format{
An object of class \code{list} of length 2.
}
\usage{
colors_moss
}
\description{
Currently includes only 'lows' and 'highs'.
}
\keyword{datasets}
