% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{get_data_db_azurecreds}
\alias{get_data_db_azurecreds}
\title{A Function to Gain Rapid Access to an Azure SQL Database}
\usage{
get_data_db_azurecreds(
  server,
  database,
  uid,
  pwd,
  auth = "sql",
  query = "SELECT 'No query given.';"
)
}
\arguments{
\item{server}{The server name for access}

\item{database}{The database name}

\item{uid}{The user ID, provided as the email added if authenticating via Azure AD or the SQL username if authenticating via SQL login}

\item{pwd}{The password, only necessary for SQL login}

\item{auth}{Either "sql" or "ad", defaults as SQL - alternative is for Azure Active Directory login}

\item{query}{The query to be run}
}
\description{
A Function to Gain Rapid Access to an Azure SQL Database
}
