% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{get_data_dl_azurekey}
\alias{get_data_dl_azurekey}
\title{A Function to Gain Rapid Access to an Azure Gen2 Data Lake}
\usage{
get_data_dl_azurekey(endpoint, key, container, directory)
}
\arguments{
\item{endpoint}{The endpoint of the data lake being accessed}

\item{key}{The access key for the endpoint}

\item{container}{The container being accesses}

\item{directory}{The directory for which we would like to collect the metadata}
}
\description{
A Function to Gain Rapid Access to an Azure Gen2 Data Lake
}
