% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmds_specific.R
\name{jmds_push_analysis}
\alias{jmds_push_analysis}
\title{A Function to Automate the Pushing of an Analysis to the JMDS estate}
\usage{
jmds_push_analysis(
  input_filename,
  output_filename,
  analysis_name,
  analysis_desc
)
}
\arguments{
\item{input_filename}{The locally held Rmd file for analysis}

\item{output_filename}{The file name of the HTML analysis output}

\item{analysis_name}{The unique name of the analysis within the database}

\item{analysis_desc}{The description of the analysis for the database}
}
\description{
A Function to Automate the Pushing of an Analysis to the JMDS estate
}
