% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmds_specific.R
\name{jmds_push_analysis_record}
\alias{jmds_push_analysis_record}
\title{A Function to Automate the Pushing of an Analysis HTML record to the JMDS estate}
\usage{
jmds_push_analysis_record(destination_file, analysis_name, analysis_desc)
}
\arguments{
\item{destination_file}{The destination file within the storage account, no filepath}

\item{analysis_name}{The unique name of the analysis for record-keeping in the database}

\item{analysis_desc}{The description of the analysis for the database}
}
\description{
A Function to Automate the Pushing of an Analysis HTML record to the JMDS estate
}
