% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmds_specific.R
\name{jmds_push_analysis_storage}
\alias{jmds_push_analysis_storage}
\title{A Function to Automate the Pushing of an Analysis HTML file to the JMDS estate}
\usage{
jmds_push_analysis_storage(file, destination_file)
}
\arguments{
\item{file}{The local analysis file}

\item{destination_file}{The destination file within the storage account, no filepath}
}
\description{
A Function to Automate the Pushing of an Analysis HTML file to the JMDS estate
}
