% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{push_blob_dl_azurekey}
\alias{push_blob_dl_azurekey}
\title{A Function to Rapidly Push a File to an Azure Gen2 Data Lake}
\usage{
push_blob_dl_azurekey(
  vault_name,
  secret,
  endpoint,
  container,
  file,
  destination_file
)
}
\arguments{
\item{vault_name}{The name of the vault (not including the full URL), just the name}

\item{secret}{The name of the secret behind which the storage account key is held}

\item{endpoint}{The endpoint of the data lake being accessed}

\item{container}{The container into which we'll push a file}

\item{file}{The name of the local file that will be published}

\item{destination_file}{The full file path of the file as it will appear in storage}
}
\description{
A Function to Rapidly Push a File to an Azure Gen2 Data Lake
}
