% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{push_data_db_azurekv}
\alias{push_data_db_azurekv}
\title{A Function to Publish a Record to an Azure SQL Database}
\usage{
push_data_db_azurekv(
  vault_name,
  uid_secret,
  pwd_secret,
  driver = "ODBC Driver 17 for SQL Server",
  server,
  database,
  table_name,
  data_frame
)
}
\arguments{
\item{vault_name}{The name of the vault (not including the full URL), just the name}

\item{uid_secret}{The name of the secret behind which the database user ID is held}

\item{pwd_secret}{The name of the secret behind which the database password is held}

\item{driver}{This parameter defaults to 'ODBC Driver 17 for SQL Server' but can be changed}

\item{server}{The Azure SQL Server into which we're depositing data}

\item{database}{The Azure SQL Database into which we're depositing data}

\item{table_name}{The name of the table into which we're depositing data}

\item{data_frame}{The data packet being passed into the table - presented as a data frame}
}
\description{
A Function to Publish a Record to an Azure SQL Database
}
